fix: try all resolved socket addrs for connection (#1651)
For the dealer connection, only the first address was tried. For example
when using a VPN, which doesn't support IPv6, the first address might
still be an IPv6 address and thus fail. The IPv4 address later in the
list is never tried. This relies on the tokio default to try all resolved
addresses and use the first that was successful.